add specs to regeneration credit
