+++ contracts/RegenerationCredit.sol
@@ -46,7 +46,9 @@ contract RegenerationCredit is ERC20, Ownable {
     address tokenOwner,
     address receiver,
     uint256 numTokens
-  ) public mustBeContractPool mustHaveRegenerationCredits(tokenOwner, numTokens) returns (bool) {
+  ) public mustBeContractPool returns (bool) {
+    require(numTokens <= balances[tokenOwner], "You don't have RC Tokens");
+
     balances[tokenOwner] = balances[tokenOwner].sub(numTokens);
     balances[receiver] = balances[receiver].add(numTokens);
     emit Transfer(tokenOwner, receiver, numTokens);
@@ -58,7 +60,7 @@ contract RegenerationCredit is ERC20, Ownable {
     return true;
   }
 
-  function contractPool(address contractFundsAddress) internal view returns (bool) {
+  function contractPool(address contractFundsAddress) public view returns (bool) {
     return contractsPools[contractFundsAddress];
   }
 
@@ -140,11 +142,6 @@ contract RegenerationCredit is ERC20, Ownable {
 
   modifier mustBeContractPool() {
     require(contractPool(msg.sender), "Not a contract pool");
-    _;
-  }
-
-  modifier mustHaveRegenerationCredits(address tokenOwner, uint256 numTokens) {
-    require(numTokens <= balances[tokenOwner], "You don't have RCT Tokens");
     _;
   }
 }
